streamlit/requirements.txt
python-dotenv
streamlit
soundfile
audio_recorder_streamlit
plotly
streamlit_extras
